Keep same package for test case classes (#33379)
