+++ features/shadow/core/src/test/java/org/apache/shardingsphere/shadow/algorithm/shadow/column/ColumnValueMatchedShadowAlgorithmTest.java
@@ -29,7 +29,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class ColumnValueMatchShadowAlgorithmTest {
+class ColumnValueMatchedShadowAlgorithmTest {
     
     private static final String SHADOW_TABLE = "t_user";
