+++ features/sharding/core/src/test/java/org/apache/shardingsphere/sharding/rewrite/token/pojo/ShardingTableTokenTest.java
@@ -35,7 +35,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.mockito.Mockito.withSettings;
 
-class TableTokenTest {
+class ShardingTableTokenTest {
     
     @Test
     void assertToString() {
