+++ features/sharding/distsql/handler/src/test/java/org/apache/shardingsphere/sharding/distsql/converter/ShardingTableRuleStatementConverterTest.java
@@ -43,7 +43,7 @@ import java.util.Properties;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-class ShardingRuleStatementConverterTest {
+class ShardingTableRuleStatementConverterTest {
     
     @Test
     void assertConvert() {
