+++ infra/database/type/clickhouse/src/test/java/org/apache/shardingsphere/infra/database/clickhouse/connector/ClickHouseConnectionPropertiesParserTest.java
@@ -38,7 +38,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class ClickHouseConnectionPropertiesTest {
+class ClickHouseConnectionPropertiesParserTest {
     
     private final ConnectionPropertiesParser parser = DatabaseTypedSPILoader.getService(ConnectionPropertiesParser.class, TypedSPILoader.getService(DatabaseType.class, "ClickHouse"));
