+++ infra/database/type/mysql/src/test/java/org/apache/shardingsphere/infra/database/mysql/checker/MySQLDatabasePrivilegeCheckerTest.java
@@ -39,7 +39,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class MySQLDatabaseEnvironmentCheckerTest {
+class MySQLDatabasePrivilegeCheckerTest {
     
     @Mock
     private PreparedStatement preparedStatement;
