+++ infra/database/type/postgresql/src/test/java/org/apache/shardingsphere/infra/database/postgresql/checker/PostgreSQLDatabasePrivilegeCheckerTest.java
@@ -39,7 +39,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class PostgreSQLDatabaseEnvironmentCheckerTest {
+class PostgreSQLDatabasePrivilegeCheckerTest {
     
     @Mock
     private DataSource dataSource;
